backend/requirements.txt
streamlit
opencv-python
numpy
ultralytics
pillow
torch
torchvision
imageio
imageio-ffmpeg
